細かい修正

+++ GameSources/ExpBar.cpp
@@ -14,7 +14,7 @@ namespace basecross {
 
 		m_bar = ObjectFactory::Create<GaugeSizeSettings>(GetStage(), 200.0f, false); 
 		m_bar->UpdateGaugeSize(m_maxExp, playerStatus->GetStatusValue(L"EXP"));
-		m_bar->SetPosition(App::GetApp()->GetGameWidth(), 0.0f, 0.46f); // 画面の左端から右端に向かってゲージを伸ばす
+		m_bar->SetPosition((float)App::GetApp()->GetGameWidth(), 0.0f, 0.46f); // 画面の左端から右端に向かってゲージを伸ばす
 	}
 
 	void ExpBar::OnUpdate()
